simrag_reproduction/requirements.txt
# Core ML/AI Libraries
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
accelerate>=0.20.0
bitsandbytes>=0.41.0

# RAG and Retrieval
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
chromadb>=0.4.0
langchain>=0.0.200
langchain-community>=0.0.10

# Evaluation and Metrics
evaluate>=0.4.0
rouge-score>=0.1.2
nltk>=3.8.1
scikit-learn>=1.3.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.65.0
pyyaml>=6.0

# Logging and Monitoring
wandb>=0.15.0
tensorboard>=2.13.0
loguru>=0.7.0

# Utilities
click>=8.1.0
rich>=13.0.0
hydra-core>=1.3.0
omegaconf>=2.3.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0
